refactor: moves logic to helpers files
